+++ src/pageComponents/HomePage/defaultData/aboutUsContent.ts
@@ -1,0 +1,31 @@
+export const aboutUsContent = {
+	about: {
+		title: 'Громадська організація «4.5.0. Кривий Ріг»',
+		subtitle: 'Офіційна неприбуткова громадська організація',
+		list: [
+			{
+				id: 'voluntary',
+				src: 'svg/big-icon--voluntary.svg',
+				text: 'Закриваємо гуманітарні потреби цивільного населення Кривого Рогу та вимушено переміщених осіб Криворізького району та Херсонської області',
+			},
+			{
+				id: 'tent',
+				src: 'svg/big-icon--tent.svg',
+				text: 'Відбудовуємо зруйновані та пошкоджені об’єкти у Криворізькому районі та Херсонській області',
+			},
+			{
+				id: 'vaccine',
+				src: 'svg/big-icon--vaccine.svg',
+				text: 'Забезпечуємо медичні заклади Кривого Рогу, населених пунктів Криворізького району та Херсонської області',
+			},
+		],
+	},
+	team: {
+		title: 'Наша команда',
+		text: 'Голова правління – Польовий Ростислав Олексійович.\nСпівзасновник організації – Кулакевич Олександр Ігорович.\nПроєкт менеджер (напрямок відбудови) – Хромих Артем Валерійович.\nПроєкт менеджер (гуманітарний напрямок) – Кулакевич Катерина Сергіївна.',
+	},
+	history: {
+		title: 'Історія заснування',
+		text: 'Громадська організація «4.5.0. Кривий Ріг» офіційно зареєстрована 27 травня 2022 року в місті Кривий Ріг. З початку своєї діяльності, ми надавали допомогу тимчасово переміщеним особам та людям з тяжкими фінансовими обставинами. З липня місяця 2022 року почали втілювати перший проєкт з відбудови. Відремонтували тимчасове укриття в Криворізькому обласному ліцеї для сільської молоді.\nЗа рік плідної праці орендуємо вже 3 склади для гуманітарної допомоги. Завдяки нашим партнерам ми продовжуємо закривати потреби цивільних та долучатись до допомоги медичним закладам міста та Херсонської області.',
+	},
+};
